refactor : mockList 위치 변경 및 코드 포맷팅
