--- WssElixir/app/src/main/java/com/example/wsselixir/data/mock/FollowerMockList.kt
@@ -1,12 +1,0 @@
-package com.example.wsselixir.data.mock
-
-import com.example.wsselixir.R
-import com.example.wsselixir.data.dataclass.Follower
-
-val followerMockList: MutableList<Follower> = mutableListOf(
-    Follower(1, "김세훈", R.drawable.ksh),
-    Follower(2, "서재원", R.drawable.sjw),
-    Follower(3, "손명지", R.drawable.smj),
-    Follower(4, "이연진", R.drawable.lyj),
-    Follower(5, "최준서", R.drawable.cjs)
-)
